refactor: 테스트 케이스 파일 분리 및 수정
- 각 상황에 맞게 UI테스트, Unit테스트 파일 분리
- beforeEach -> before로 수정 가능한 부분 수정
